apzkr-pzpi-21-8-holia-stanislav/Task1-Server/fieldforge/djangoapp/serializers/results_serializer.py
from rest_framework import serializers # type: ignore
from djangoapp.models import Results

class ResultsSerializer(serializers.ModelSerializer):
    class Meta:
        model = Results
        fields = '__all__'
        read_only_fields = ['uuid']