apzkr-pzpi-21-8-holia-stanislav/Task1-Server/fieldforge/requirements.txt
Django==3.2.5
psycopg2-binary==2.9.1
djangorestframework==3.12.4
pytz==2021.1